Minor clean up
